smth

+++ Entities/Finances.cs
@@ -3,6 +3,6 @@ namespace Aktan.Entities
 public class Finances
 {
     public int FinancesId{get;set;}
-    public float Sum{get;set;}
+    public double Sum{get;set;}
 }
 }
